+++ packages/usage-pricing/src/pricing/pricing.ts
@@ -2,6 +2,7 @@ import { Usage } from '@microfox/usage-tracker';
 import { OpenAIPricingConfig } from './ai-provider-openai';
 import { AwsSesPricingConfig } from './aws-ses';
 import { BravePricingConfig } from './brave';
+import { UsageWithPricing } from '../types';
 
 export const PricingConfig = {
   ...AwsSesPricingConfig,
@@ -9,40 +10,58 @@ export const PricingConfig = {
   ...BravePricingConfig,
 };
 
-export const attachPricingApi1 = (usage: Usage) => {
-  const pricingConfig = (PricingConfig as any)[usage.package];
-  if (!pricingConfig) {
-    return 0;
+export const attachPricingApi1 = (
+  usage: UsageWithPricing,
+): UsageWithPricing => {
+  if (usage.type !== 'api_1' || !usage.requestKey) {
+    return usage;
+  }
+  const pricingPkgConfig = (PricingConfig as any)[usage.package];
+  if (!pricingPkgConfig) {
+    return {
+      ...usage,
+      priceUSD: 0,
+    };
+  }
+  const pricingKeyConfig = (pricingPkgConfig as any)[usage.requestKey];
+  if (!pricingKeyConfig) {
+    return {
+      ...usage,
+      priceUSD: 0,
+    };
   }
   let usagePriceUSD = 0;
   if (
     usage.type === 'api_1' &&
     usage.requestCount &&
-    pricingConfig.requestCount
+    pricingKeyConfig.requestCount
   ) {
     usagePriceUSD +=
-      pricingConfig.requestCount.basePriceUSD *
-      (usage.requestCount / pricingConfig.requestCount.per);
+      pricingKeyConfig.requestCount.basePriceUSD *
+      (usage.requestCount / pricingKeyConfig.requestCount.per);
   }
   if (
     usage.type === 'api_1' &&
     usage.requestData &&
-    pricingConfig.requestData
+    pricingKeyConfig.requestData
   ) {
     usagePriceUSD +=
-      pricingConfig.requestData.basePriceUSD *
-      (usage.requestData / pricingConfig.requestData.per);
+      pricingKeyConfig.requestData.basePriceUSD *
+      (usage.requestData / pricingKeyConfig.requestData.per);
   }
   return {
     ...usage,
     priceUSD: usagePriceUSD,
-  };
+  } as UsageWithPricing;
 };
 
-export const attachPricingLLM = (usage: Usage) => {
+export const attachPricingLLM = (usage: UsageWithPricing): UsageWithPricing => {
   const pricingConfig = (PricingConfig as any)[usage.package];
   if (!pricingConfig) {
-    return 0;
+    return {
+      ...usage,
+      priceUSD: 0,
+    };
   }
   let usagePriceUSD = 0;
   if (usage.type != 'llm' || !usage.model) {
@@ -64,11 +83,12 @@ export const attachPricingLLM = (usage: Usage) => {
   };
 };
 
-export const attachPricing = (usage: Usage) => {
+export const attachPricing = (usage: UsageWithPricing): UsageWithPricing => {
   if (usage.type === 'llm') {
     return attachPricingLLM(usage);
   }
   if (usage.type === 'api_1') {
     return attachPricingApi1(usage);
   }
+  return usage;
 };