+++ packages/agent-kit/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @microfox/webhook-slack
 
+## 0.0.3
+
+### Patch Changes
+
+- 9c9a795: Changes from PR #362: updates
+- Updated dependencies [9c9a795]
+  - @microfox/webhook-slack@1.0.9
+
 ## 0.0.2
 
 ### Patch Changes