+++ packages/brave/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @microfox/brave
 
+## 1.0.10
+
+### Patch Changes
+
+- 82d5fb7: Changes from PR #370: updates
+- Updated dependencies [82d5fb7]
+  - @microfox/usage-tracker@0.0.3
+
 ## 1.0.9
 
 ### Patch Changes